Transform_ROI
This script will transform a ROI created in standard space to native space. It transforms the native space clean nifti to standard space and records a .mat file with the distortions to change the clean nifti to standard space, and applies those calculations to the ROI to transform it from standard to native space.